test: move cold setup outside test budgets
